pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robin-osint"
version = "0.1.0"
description = "AI-Powered Dark Web OSINT Tool"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Apurv Singh Gautam"}
]
keywords = ["osint", "darkweb", "tor", "investigation", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "requests>=2.32.3",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3.0",
    "PySocks>=1.7.1",
    "googletrans==4.0.0rc1",
    "langchain>=0.3.14",
    "langchain-openai>=0.2.14",
    "langchain-anthropic>=0.3.3",
    "langchain-google-genai>=2.0.8",
    "langchain-community>=0.3.13",
    "python-dotenv>=1.0.1",
    "croniter>=5.0.1",
    "pyyaml>=6.0.0",
    "truffleHog3>=3.0.0",
    "h8mail>=2.5.6",
    "credentialdigger>=4.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
robin = "robin.cli:robin"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
